requirements.txt
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
django-jazzmin==3.0.4
pillow==12.2.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2026.1
